refactor: refactor attention hint decoration implement
