+++ spx-gui/src/components/editor/code-editor/coordinators/index.ts
@@ -8,14 +8,14 @@ import {
   type InlayHintDecoration,
   type InputItem,
   type InputItemCategory,
-  type LayerContent,
   type SelectionMenuItem,
   type TextModel
 } from '@/components/editor/code-editor/EditorUI'
 import { Runtime } from '../runtime'
+import type { Definition } from '../compiler'
 import { Compiler } from '../compiler'
 import { ChatBot } from '../chat-bot'
-import { DocAbility, type Doc } from '../document'
+import { DocAbility } from '../document'
 import { Project } from '@/models/project'
 import { type IRange, type Position } from 'monaco-editor'
 import type { I18n } from '@/utils/i18n'
@@ -34,248 +34,12 @@ import {
   TokenType
 } from '@/components/editor/code-editor/tools'
 import { debounce } from '@/utils/utils'
-import type { Definition, DefinitionUsage, Diagnostic } from '../compiler'
+import { HoverProvider } from '@/components/editor/code-editor/coordinators/hoverProvider'
 
 type JumpPosition = {
   line: number
   column: number
   fileUri: string
-}
-
-const VariableDefinitionType = [
-  'bool',
-  'int',
-  'int8',
-  'int16',
-  'int32',
-  'int64',
-  'uint',
-  'uint8',
-  'uint16',
-  'uint32',
-  'uint64',
-  'uintptr',
-  'float32',
-  'float64',
-  'complex64',
-  'complex128',
-  'string',
-  'byte',
-  'rune'
-]
-
-type CoordinatorState = {
-  definitions: Definition[]
-  diagnostics: Diagnostic[]
-}
-
-class HoverProvider {
-  private ui: EditorUI
-  private docAbility: DocAbility
-  private state: CoordinatorState
-
-  constructor(ui: EditorUI, docAbility: DocAbility, state: CoordinatorState) {
-    this.ui = ui
-    this.docAbility = docAbility
-    this.state = state
-  }
-
-  async provideHover(
-    _model: TextModel,
-    ctx: {
-      position: Position
-      hoverUnitWord: string
-      signal: AbortSignal
-    }
-  ): Promise<LayerContent[]> {
-    const contents = this.docAbility.getNormalDoc({
-      module: '',
-      name: ctx.hoverUnitWord
-    })
-
-    const definition = this.findDefinition(ctx.position)
-    const diagnostic = this.findDiagnostic(ctx.position.lineNumber)
-    const matchedContent = this.findMatchedContent(contents, definition)
-    const layerContents: LayerContent[] = []
-
-    if (diagnostic) layerContents.push(this.createDiagnosticContent(diagnostic))
-
-    if (contents && contents.length > 0) {
-      layerContents.push(...this.createDocContents(contents))
-    } else if (definition && definition.usages.length > 0) {
-      // in definition, usages have only one usage
-      const [usage] = definition.usages
-      const content = VariableDefinitionType.includes(usage.type)
-        ? this.createVariableRenameContent(usage)
-        : this.createDefinitionContent(usage, matchedContent)
-      layerContents.push(content)
-    }
-
-    return layerContents
-  }
-
-  private findDefinition(position: Position): Definition | undefined {
-    return this.state.definitions.find((def) => {
-      const tokenLen = def.end_pos - def.start_pos
-      const line = def.start_position.Line
-      const startColumn = def.start_position.Column
-      const endColumn = startColumn + tokenLen
-      return (
-        position.lineNumber === line &&
-        position.column >= startColumn &&
-        position.column <= endColumn
-      )
-    })
-  }
-
-  private findDiagnostic(lineNumber: number): Diagnostic | undefined {
-    return this.state.diagnostics.find((diag) => diag.line === lineNumber)
-  }
-
-  private findMatchedContent(contents: Doc[] | null, definition: Definition | undefined) {
-    return contents?.find(
-      (content) =>
-        definition &&
-        content.token.id.name === definition.pkg_name &&
-        content.token.id.module === definition.pkg_path
-    )
-  }
-
-  private createDiagnosticContent(diagnostic: Diagnostic): LayerContent {
-    return {
-      type: 'doc',
-      layer: {
-        level: DocPreviewLevel.Error,
-        content: diagnostic.message,
-        recommendAction: {
-          label: this.ui.i18n.t({
-            zh: '没看明白？场外求助',
-            en: 'Not clear? Ask for help'
-          }),
-          activeLabel: this.ui.i18n.t({ zh: '在线答疑', en: 'Online Q&A' }),
-          onActiveLabelClick: () => {
-            // TODO: Add some logic code
-          }
-        }
-      }
-    }
-  }
-
-  private createVariableRenameContent(usage: DefinitionUsage): LayerContent {
-    return {
-      type: 'doc',
-      layer: {
-        level: DocPreviewLevel.Normal,
-        header: {
-          icon: Icon.Variable,
-          declaration: usage.declaration
-        },
-        content: '',
-        moreActions: [
-          {
-            icon: Icon.Rename,
-            label: this.ui.i18n.t({ zh: '重命名', en: 'Rename' }),
-            onClick: () => ({
-              type: 'rename',
-              layer: {
-                placeholder: this.ui.i18n.t({
-                  zh: '请输入新名称',
-                  en: 'Please enter a new name'
-                }),
-                onSubmit: async (
-                  newName: string,
-                  ctx: { signal: AbortSignal },
-                  setError: (message: string) => void
-                ) => {
-                  // TODO: Add some logic code
-                }
-              }
-            })
-          }
-        ]
-      }
-    }
-  }
-
-  private createDefinitionContent(
-    usage: DefinitionUsage,
-    matchedContent: Doc | undefined
-  ): LayerContent {
-    const actions = matchedContent ? this.createActions(matchedContent) : {}
-    return {
-      type: 'doc',
-      layer: {
-        level: DocPreviewLevel.Normal,
-        content: matchedContent?.content,
-        header: {
-          icon: Icon.Function,
-          declaration: usage.declaration
-        },
-        ...actions
-      }
-    }
-  }
-
-  private createActions(matchedContent: Doc) {
-    return {
-      recommendAction: {
-        label: this.ui.i18n.t({
-          zh: '还有疑惑？场外求助',
-          en: 'Still in confusion? Ask for help'
-        }),
-        activeLabel: this.ui.i18n.t({ zh: '在线答疑', en: 'Online Q&A' }),
-        onActiveLabelClick: () => {
-          // TODO: Add some logic code
-        }
-      },
-      moreActions: [
-        {
-          icon: Icon.Document,
-          label: this.ui.i18n.t({ zh: '查看文档', en: 'Document' }),
-          onClick: () => {
-            const detailDoc = this.docAbility.getDetailDoc(matchedContent.token)
-            if (!detailDoc) return
-            this.ui.invokeDocumentDetail(detailDoc.content)
-          }
-        }
-      ]
-    }
-  }
-
-  private createDocContents(contents: any[]): LayerContent[] {
-    return contents.map((doc) => ({
-      type: 'doc',
-      layer: {
-        level: DocPreviewLevel.Normal,
-        header: {
-          icon: Icon.Function,
-          declaration: '' // TODO: implement document struct and set declaration
-        },
-        content: doc.content,
-        recommendAction: {
-          label: this.ui.i18n.t({
-            zh: '还有疑惑？场外求助',
-            en: 'Still in confusion? Ask for help'
-          }),
-          activeLabel: this.ui.i18n.t({ zh: '在线答疑', en: 'Online Q&A' }),
-          onActiveLabelClick: () => {
-            // TODO: add some logic code here
-          }
-        },
-        moreActions: [
-          {
-            icon: Icon.Document,
-            label: this.ui.i18n.t({ zh: '查看文档', en: 'Document' }),
-            onClick: () => {
-              const detailDoc = this.docAbility.getDetailDoc(doc.token)
-              if (!detailDoc) return
-              this.ui.invokeDocumentDetail(detailDoc.content)
-            }
-          }
-        ]
-      }
-    }))
-  }
 }
 
 export class Coordinator {
@@ -285,10 +49,7 @@ export class Coordinator {
   docAbility: DocAbility
   compiler: Compiler
   public updateDefinition = debounce(this._updateDefinition, 300)
-  private state: CoordinatorState = {
-    definitions: [],
-    diagnostics: []
-  }
+  private definitions: Definition[] = []
   private hoverProvider: HoverProvider
 
   constructor(
@@ -304,7 +65,7 @@ export class Coordinator {
     this.docAbility = docAbility
     this.chatBot = chatBot
     this.compiler = compiler
-    this.hoverProvider = new HoverProvider(ui, docAbility, this.state)
+    this.hoverProvider = new HoverProvider(ui, docAbility, this.definitions)
 
     ui.registerCompletionProvider({
       provideDynamicCompletionItems: this.implementsPreDefinedCompletionProvider.bind(this)
@@ -503,7 +264,6 @@ export class Coordinator {
       .getDiagnostics(this.currentFilename, this.getProjectAllCodes())
       .then((diagnostics) => {
         if (ctx.signal.aborted) return
-        this.state.diagnostics = diagnostics
         setHints(
           diagnostics.map((diagnostic) => {
             const word = model.getWordAtPosition({
@@ -524,7 +284,17 @@ export class Coordinator {
                 type: 'doc',
                 layer: {
                   level: DocPreviewLevel.Error,
-                  content: diagnostic.message
+                  content: diagnostic.message,
+                  recommendAction: {
+                    label: this.ui.i18n.t({
+                      zh: '没看明白？场外求助',
+                      en: 'Not clear? Ask for help'
+                    }),
+                    activeLabel: this.ui.i18n.t({ zh: '在线答疑', en: 'Online Q&A' }),
+                    onActiveLabelClick: () => {
+                      // TODO: Add some logic code
+                    }
+                  }
                 }
               }
             }
@@ -551,11 +321,10 @@ export class Coordinator {
   }
 
   private async _updateDefinition() {
-    const definition = await this.compiler.getDefinition(
+    this.definitions = await this.compiler.getDefinition(
       this.currentFilename,
       this.getProjectAllCodes()
     )
-    this.state.definitions = definition
   }
 
   public jump(position: JumpPosition): void {}
